Added viewers and the ability to add and remove users from your catalog.

+++ backend/api/v1/viewer.php
@@ -12,13 +12,14 @@ $app->group('/api', function () use ($app) {
         * GET
         * ========================================================== */
 
-        /* Get all viewers */
-        $app->get($resource, function ($request, $response, $args) use ($app)
+        /* Get all viewers for a creator for a specific status (approved, pending, rejected) */
+        $app->get($resource . "/list/{status}", function ($request, $response, $args) use ($app)
         {
             $session = APIService::authenticate_request($_GET);
             $user_id = $session->user->id;
+            $status = $args['status'];
 
-            $viewers = Viewer::get_all($user_id);
+            $viewers = Viewer::get_all($user_id, $status);
             if($viewers === false) {
                 APIService::response_fail("There was a problem getting the viewers.", 500);
             }
@@ -34,9 +35,10 @@ $app->group('/api', function () use ($app) {
 
             $creator_id = intval($args['creator_id']);
             $viewer_id = $session->user->id;
+            $status = "approved";
 
-            $viewer = Viewer::get_for_creator_and_viewer_id($creator_id, $viewer_id);
-            if($viewer === false) {
+            $viewer = Viewer::get_for_creator_and_viewer_id($creator_id, $viewer_id, $status);
+            if($viewer === false || $viewer === null) {
                 APIService::response_fail("There was a problem getting viewer.", 500);
             }
             if($viewer === null) {
@@ -46,12 +48,13 @@ $app->group('/api', function () use ($app) {
         });
 
         /* Get all creator can view */
-        $app->get($resource . "/view/list", function ($request, $response, $args) use ($app)
+        $app->get($resource . "/view/list/{status}", function ($request, $response, $args) use ($app)
         {
             $session = APIService::authenticate_request($_GET);
             $user_id = $session->user->id;
+            $status = $args['status'];
 
-            $viewers = Viewer::get_all_user_views($user_id);
+            $viewers = Viewer::get_all_user_views($user_id, $status);
             if($viewers === false) {
                 APIService::response_fail("There was a problem getting the viewers.", 500);
             }
@@ -71,9 +74,17 @@ $app->group('/api', function () use ($app) {
             $user_id = $session->user->id;
 
             $params = APIService::build_params($_REQUEST, array(
+                "creator_id",
                 "viewer_id"
             ), array());
-            $params["creator_id"] = $user_id;
+
+            if ($user_id == $params["creator_id"]){
+                $params["status"] = "approved";
+            }else if ($user_id == $params["viewer_id"]){
+                $params["status"] = "pending";
+            }else{
+                APIService::response_fail("There was a problem creating the viewer.", 500);
+            }
 
             $viewer = Viewer::create_from_data($params);
             if($viewer === false || $viewer === null) {
@@ -83,20 +94,52 @@ $app->group('/api', function () use ($app) {
         });
 
         /* ========================================================== *
+        * PUT
+        * ========================================================== */
+
+        /* Approve or reject a viewer */
+        $app->post($resource . '/{id}', function ($request, $response, $args) use ($app)
+        {
+            $session = APIService::authenticate_request($_REQUEST);
+            $user_id = $session->user->id;
+            $id = intval($args["id"]);
+
+            $params = APIService::build_params($_REQUEST, null, array(
+                "status"
+            ));
+
+            $status = $params["status"];
+            if ($status !== "approved" && $status !== "rejected" && $status !== "pending"){
+                APIService::response_fail("Invalid status.", 500);
+            }
+            
+            $viewer = Viewer::update($user_id, $id, $params);
+            if($viewer === false || $viewer === null) {
+                APIService::response_fail("There was a problem updating the viewer.", 500);
+            }
+            APIService::response_success($viewer);
+        });
+
+        /* ========================================================== *
         * DELETE
         * ========================================================== */
 
         /* Delete a viewer */
-        $app->delete($resource . '/{viewer_id}', function ($response, $request, $args) use ($app)
+        $app->delete($resource . '/{viewer_id}/{creator_id}', function ($response, $request, $args) use ($app)
         {
             $session = APIService::authenticate_request($_GET);
-            $creator_id = $session->user->id;
-            $viewer_id = intval($args['viewer_id']);
+            $user_id = $session->user->id;
+            $creator_id = intval($args["creator_id"]);
+            $viewer_id = intval($args["viewer_id"]);
 
-            $result = Viewer::set_active($creator_id, $viewer_id, 0);
+            if ($user_id !== $creator_id && $user_id !== $viewer_id){
+                APIService::response_fail("There was a problem deleting the viewer.", 500);
+            }
+
+            $result = Viewer::delete($creator_id, $viewer_id);
 
             if( $result === false ) {
-                APIService::response_fail("There was an error setting the active state of that viewer.", 500);
+                APIService::response_fail("There was an error deleting that viewer.", 500);
             }
             if( $result === null ) {
                 APIService::response_fail("The requested viewer does not exist.", 404);
